requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.0
sqlalchemy==2.0.30
sqlalchemy-utils==0.41.1
pymysql==1.1.0
starlette>=0.40.0,<0.42.0
pydantic==2.6.4
itsdangerous==2.2.0
python-dotenv